requirements.txt
Django>=1.11,<2.0
djangorestframework
git+https://github.com/anx-ckreuzberger/djangodav
